+++ packages/set/src/venn/VennIntersectionLabels.tsx
@@ -1,14 +1,16 @@
 import { createElement, ReactNode } from 'react'
-import { getClassName, Label, NumericPositionSpec, X, Y } from '@chsk/core'
+import { getClassName, Label } from '@chsk/core'
 import { isVennPreparedData } from './predicates'
 import { VennIntersectionLabelsProps, VennInteractiveDataItem } from './types'
 import { useVennPreparedData } from './contexts'
 
 export const VennIntersectionLabels = ({
+    ids,
     format = (v: string | number, item?: VennInteractiveDataItem) =>
         String(v) ?? String(item?.value),
     padding = [4, 4, 4, 4],
     size = [40, 10],
+    anchor = [0.5, 0.5],
     align = [0.5, 0.5],
     offset = [0, 0],
     className,
@@ -20,27 +22,24 @@ export const VennIntersectionLabels = ({
     const data = preparedData.data
     if (!isVennPreparedData(data)) return null
 
+    const idSet = new Set(ids ?? preparedData.data.map(item => item.id))
     const compositeClassName = getClassName('vennIntersectionLabel', className)
-    const result: Array<ReactNode> = data.map((item, i) => {
-        const labelPosition: NumericPositionSpec = [
-            item.labelPosition[X] + offset[X],
-            item.labelPosition[Y] + offset[Y],
-        ]
-        return createElement(
-            component,
-            {
-                key: 'venn-intersection-label-' + i,
-                position: labelPosition,
-                size,
-                align,
-                padding,
-                className: compositeClassName,
-                style,
-                setRole,
-            },
-            format(item.value, item)
-        )
-    })
+    const componentProps = { size, anchor, align, offset, padding, style, setRole }
+    const result: Array<ReactNode> = data
+        .map((item, i) => {
+            if (!idSet.has(item.id)) return null
+            return createElement(
+                component,
+                {
+                    key: 'intersection-' + i,
+                    position: item.labelPosition,
+                    className: compositeClassName,
+                    ...componentProps,
+                },
+                format(item.value, item)
+            )
+        })
+        .filter(Boolean)
 
     return <>{result}</>
 }